+++ bot/src/main/java/edu/java/bot/api/controllers/UpdatesApi.java
@@ -6,8 +6,8 @@
 
 package edu.java.bot.api.controllers;
 
-import edu.java.bot.api.models.ApiErrorResponse;
-import edu.java.bot.api.models.LinkUpdate;
+import api.bot.models.LinkUpdate;
+import api.models.ApiErrorResponse;
 import edu.java.bot.api.services.UpdatesApiService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
