--- bot/src/main/java/edu/java/bot/clients/scrapper/model/LinkResponse.java
@@ -1,105 +1,0 @@
-package edu.java.bot.clients.scrapper.model;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import io.swagger.v3.oas.annotations.media.Schema;
-import java.net.URI;
-import java.util.Objects;
-import javax.validation.Valid;
-
-/**
- * LinkResponse
- */
-
-public class LinkResponse {
-
-    private Long id;
-
-    private URI url;
-
-    public LinkResponse(Long id, URI url) {
-        this.id = id;
-        this.url = url;
-    }
-
-    public LinkResponse id(Long id) {
-        this.id = id;
-        return this;
-    }
-
-    /**
-     * Get id
-     *
-     * @return id
-     */
-
-    @Schema(name = "id", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
-    @JsonProperty("id")
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public LinkResponse url(URI url) {
-        this.url = url;
-        return this;
-    }
-
-    /**
-     * Get url
-     *
-     * @return url
-     */
-    @Valid
-    @Schema(name = "url", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
-    @JsonProperty("url")
-    public URI getUrl() {
-        return url;
-    }
-
-    public void setUrl(URI url) {
-        this.url = url;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        LinkResponse linkResponse = (LinkResponse) o;
-        return Objects.equals(this.id, linkResponse.id)
-            && Objects.equals(this.url, linkResponse.url);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, url);
-    }
-
-    @Override
-    public String toString() {
-        StringBuilder sb = new StringBuilder();
-        sb.append("class LinkResponse {\n");
-        sb.append("    id: ").append(toIndentedString(id)).append("\n");
-        sb.append("    url: ").append(toIndentedString(url)).append("\n");
-        sb.append("}");
-        return sb.toString();
-    }
-
-    /**
-     * Convert the given object to string with each line indented by 4 spaces
-     * (except the first line).
-     */
-    private String toIndentedString(Object o) {
-        if (o == null) {
-            return "null";
-        }
-        return o.toString().replace("\n", "\n    ");
-    }
-}
-
