+++ common/src/main/java/api/bot/models/LinkUpdate.java
@@ -1,4 +1,4 @@
-package edu.java.bot.api.models;
+package api.bot.models;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -22,6 +22,13 @@ public class LinkUpdate {
 
     @Valid
     private List<Long> tgChatIds;
+
+    public LinkUpdate(Long id, URI url, String description, List<Long> tgChatIds) {
+        this.id = id;
+        this.url = url;
+        this.description = description;
+        this.tgChatIds = tgChatIds;
+    }
 
     public LinkUpdate id(Long id) {
         this.id = id;
