+++ common/src/main/java/api/models/ApiErrorResponse.java
@@ -1,4 +1,4 @@
-package edu.java.bot.api.models;
+package api.models;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
