+++ scrapper/src/main/java/edu/java/api/handlers/ApiExceptionHandler.java
@@ -1,9 +1,9 @@
 package edu.java.api.handlers;
 
+import api.models.ApiErrorResponse;
 import edu.java.api.exceptions.base.BadRequestException;
 import edu.java.api.exceptions.base.ConflictException;
 import edu.java.api.exceptions.base.NotFoundException;
-import edu.java.api.models.ApiErrorResponse;
 import java.util.Arrays;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
