+++ scrapper/src/main/java/edu/java/clients/bot/BotClient.java
@@ -1,6 +1,6 @@
 package edu.java.clients.bot;
 
-import edu.java.clients.bot.models.LinkUpdate;
+import api.bot.models.LinkUpdate;
 
 public interface BotClient {
