+++ scrapper/src/main/java/edu/java/clients/github/GitHubClient.java
@@ -1,4 +1,6 @@
-package edu.java.github;
+package edu.java.clients.github;
+
+import edu.java.clients.github.models.RepositoryResponse;
 
 public interface GitHubClient {
     RepositoryResponse getRepositoryResponse(String owner, String repo);
