+++ scrapper/src/main/java/edu/java/clients/stackoverflow/StackOverflowClient.java
@@ -1,4 +1,6 @@
-package edu.java.stackoverflow;
+package edu.java.clients.stackoverflow;
+
+import edu.java.clients.stackoverflow.models.QuestionResponse;
 
 public interface StackOverflowClient {
     QuestionResponse getQuestionResponse(Long questionId);
