+++ scrapper/src/main/java/edu/java/clients/stackoverflow/models/QuestionResponse.java
@@ -1,4 +1,4 @@
-package edu.java.stackoverflow;
+package edu.java.clients.stackoverflow.models;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.time.OffsetDateTime;
