+++ scrapper/src/main/java/edu/java/models/Link.java
@@ -4,7 +4,6 @@ import java.net.URI;
 
 public record Link(
     Long id,
-
     URI url
 ) {
 }
