+++ scrapper/src/test/java/edu/scrapper/stackoverflow/StackOverflowClientTest.java
@@ -4,8 +4,8 @@ import com.github.tomakehurst.wiremock.client.WireMock;
 import com.github.tomakehurst.wiremock.junit5.WireMockExtension;
 import com.github.tomakehurst.wiremock.junit5.WireMockTest;
 import edu.java.ScrapperApplication;
-import edu.java.stackoverflow.QuestionResponse;
-import edu.java.stackoverflow.StackOverflowClient;
+import edu.java.clients.stackoverflow.models.QuestionResponse;
+import edu.java.clients.stackoverflow.StackOverflowClient;
 import java.io.File;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
